move lucene search into pyserini.search.lucene (#991)
* moved files from pyserini.search to pyserini.search.lucene.  

* change SimpleSearcher etc. to LuceneSearcher

* create fakes class and interface to make old API call still work.
